[fix] Fixes per PR review: removed tests/unit/unit_tests_tensor.cpp. Added common alloc test overloading to rmalloc.h
